scrap-service/requirements.txt
scrapy==2.12.0
beautifulsoup4==4.9.3
twisted==24.11.0
psycopg2-binary==2.9.3
